feat: create version page and manage users for version

+++ src/pages/app/p/[p]/channel/[channel].vue
@@ -1,31 +1,39 @@
 <script setup lang="ts">
 import {
   IonButton, IonButtons, IonContent,
-  IonFab,
-  IonFabButton,
   IonHeader, IonIcon, IonInput, IonItem, IonItemDivider, IonLabel, IonList,
   IonPage, IonTitle, IonToolbar, actionSheetController,
   toastController,
 } from '@ionic/vue'
-import { add, chevronBack } from 'ionicons/icons'
+import { chevronBack } from 'ionicons/icons'
 import copy from 'copy-to-clipboard'
 import { useSupabase } from '~/services/supabase'
 import type { definitions } from '~/types/supabase'
+import { openVersion } from '~/services/versions'
 
+interface ChannelUsers {
+  user_id: definitions['users']
+}
+interface Channel {
+  version: definitions['app_versions']
+}
 const { t } = useI18n()
 const router = useRouter()
 const route = useRoute()
 const supabase = useSupabase()
 const packageId = ref<string>('')
 const id = ref<number>()
-const channel = ref<definitions['channels']>()
-const users = ref<definitions['channel_users']>()
+const channel = ref<definitions['channels'] & Channel>()
+const users = ref<(definitions['channel_users'] & ChannelUsers)[]>()
 const newUser = ref<string>()
 
-interface ChannelUsers {
-  users_id: definitions['users'][]
+const openApp = () => {
+  if (!channel.value) return
+  openVersion(channel.value.version)
 }
 const getUsers = async() => {
+  if (!channel.value)
+    return
   try {
     const { data: dataUsers } = await supabase
       .from<definitions['channel_users'] & ChannelUsers>('channel_users')
@@ -33,6 +41,7 @@ const getUsers = async() => {
           id,
           channel_id,
           user_id (
+            id,
             email,
             first_name,
             last_name
@@ -40,11 +49,39 @@ const getUsers = async() => {
           created_at
         `)
       .eq('channel_id', id.value)
-      .eq('app_id', channel.value.app_id)
+      .eq('app_id', channel.value.version.app_id)
     if (dataUsers && dataUsers.length)
       users.value = dataUsers
     else
       console.log('no users')
+    console.log('users', users.value)
+  }
+  catch (error) {
+    console.error(error)
+  }
+}
+const getChannel = async() => {
+  try {
+    const { data } = await supabase
+      .from<definitions['channels'] & Channel>('channels')
+      .select(`
+          id,
+          name,
+          version (
+            name,
+            app_id,
+            bucket_id,
+            created_at
+          ),
+          created_at,
+          updated_at
+        `)
+      .eq('id', id.value)
+    if (data && data.length)
+      channel.value = data[0]
+    else
+      console.log('no channel')
+    console.log('channel', channel.value)
   }
   catch (error) {
     console.error(error)
@@ -55,24 +92,14 @@ watchEffect(async() => {
     packageId.value = route.params.p as string
     packageId.value = packageId.value.replaceAll('--', '.')
     id.value = Number(route.params.channel as string)
-    try {
-      const { data: dataApp } = await supabase
-        .from<definitions['channels']>('channels')
-        .select()
-        .eq('id', id.value)
-      if (dataApp && dataApp.length)
-        channel.value = dataApp[0]
-      else
-        console.log('no channel')
-      await getUsers()
-    }
-    catch (error) {
-      console.error(error)
-    }
+    await getChannel()
+    await getUsers()
   }
 })
 const addUser = async() => {
   console.log('newUser', newUser.value)
+  if (!channel.value)
+    return
   const { data, error: uError } = await supabase
     .from<definitions['users']>('users')
     .select()
@@ -81,7 +108,7 @@ const addUser = async() => {
     console.log('no user', uError)
     const toast = await toastController
       .create({
-        message: 'User not found, ask to register first',
+        message: t('channel.user_no_found'),
         duration: 2000,
       })
     await toast.present()
@@ -92,7 +119,7 @@ const addUser = async() => {
     .from<definitions['channel_users']>('channel_users')
     .insert({
       channel_id: id.value,
-      app_id: channel.value.app_id,
+      app_id: channel.value.version.app_id,
       user_id: data[0].id,
     })
   if (error) console.error(error)
@@ -106,29 +133,33 @@ const publicLink = computed(() => channel.value ? `https://capgp.app/api/latest?
 const copyPublicLink = () => {
   copy(publicLink.value)
 }
-const makePublic = async() => {
+const makePublic = async(val = true) => {
+  if (!channel.value)
+    return
   const { error } = await supabase
     .from<definitions['channels']>('channels')
-    .update({ public: true })
+    .update({ public: val })
     .eq('id', id.value)
   if (error) {
     console.error(error)
   }
   else {
-    channel.value.public = true
+    channel.value.public = val
     const toast = await toastController
       .create({
-        message: 'Defined as public',
+        message: `Defined as ${val ? 'public' : 'private'}`,
         duration: 2000,
       })
     await toast.present()
   }
 }
 const deleteUser = async(usr: definitions['users']) => {
+  if (!channel.value)
+    return
   const { error } = await supabase
     .from<definitions['channel_users']>('channel_users')
     .delete()
-    .eq('app_id', channel.value.app_id)
+    .eq('app_id', channel.value.version.app_id)
     .eq('user_id', usr.id)
   if (error) console.error(error)
   else
@@ -162,26 +193,22 @@ const presentActionSheet = async(usr: definitions['users']) => {
       <IonToolbar class="toolbar-no-border">
         <IonButtons slot="start" class="mx-3">
           <IonButton @click="back">
-            <IonIcon :icon="chevronBack" class="text-grey-dark" />
+            <IonIcon :icon="chevronBack" class="text-grey-dark" /> {{ t('button.back') }}
           </IonButton>
         </IonButtons>
-        <div class="flex justify-between items-center">
-          <div class="flex">
-            <div class="flex flex-col justify-center">
-              <p class="text-left text-bright-cerulean-500 text-sm font-bold">
-                {{ channel?.name }}
-              </p>
-            </div>
-          </div>
-        </div>
       </IonToolbar>
     </IonHeader>
     <ion-content :fullscreen="true">
       <ion-header>
         <ion-toolbar>
           <ion-title size="large">
-            {{ t('channel.title') }}
+            {{ t('channel.title') }} {{ channel?.name }}
           </ion-title>
+          <IonButtons v-if="channel" slot="end">
+            <IonButton color="danger" @click="openApp()">
+              {{ t('channel.open') }}
+            </IonButton>
+          </IonButtons>
         </ion-toolbar>
       </ion-header>
       <ion-list>
@@ -201,12 +228,23 @@ const presentActionSheet = async(usr: definitions['users']) => {
             </div>
           </IonLabel>
         </IonItem>
+        <IonItem v-if="channel && channel.public" @click="makePublic(false)">
+          <IonLabel>
+            <div class="col-span-6 flex flex-col cursor-pointer">
+              <div class="flex justify-between items-center">
+                <h2 class="text-sm text-bright-cerulean-500">
+                  {{ t('channel.makeprivate') }}
+                </h2>
+              </div>
+            </div>
+          </IonLabel>
+        </IonItem>
         <IonItem v-else @click="makePublic()">
           <IonLabel>
             <div class="col-span-6 flex flex-col">
               <div class="flex justify-between items-center cursor-pointer">
                 <h2 class="text-sm text-bright-cerulean-500">
-                  Your channel is private, click to make it public
+                  {{ t('channel.makepublic') }}
                 </h2>
               </div>
             </div>
@@ -219,11 +257,14 @@ const presentActionSheet = async(usr: definitions['users']) => {
         </ion-item-divider>
         <ion-item>
           <ion-label position="floating">
-            Email address to invite
+            {{ t('channel.invit') }}
           </ion-label>
-          <ion-input v-model="newUser" type="email" placeholder="Enter Email" />
+          <ion-input v-model="newUser" type="email" placeholder="hello@yourcompany.com" />
+          <ion-button slot="end" @click="addUser()">
+            {{ t('channel.add') }}
+          </ion-button>
         </ion-item>
-        <IonItem v-for="(usr, index) in users" :key="index" @click="presentActionSheet(usr)">
+        <IonItem v-for="(usr, index) in users" :key="index" @click="presentActionSheet(usr.user_id)">
           <IonLabel>
             <div class="col-span-6 flex flex-col">
               <div class="flex justify-between items-center">
@@ -236,11 +277,6 @@ const presentActionSheet = async(usr: definitions['users']) => {
           </IonLabel>
         </IonItem>
       </ion-list>
-      <ion-fab slot="fixed" horizontal="end" vertical="bottom">
-        <ion-fab-button color="danger" @click="addUser()">
-          <IonIcon :icon="add" />
-        </ion-fab-button>
-      </ion-fab>
     </ion-content>
   </ion-page>
 </template>
